requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0
boto3>=1.34.0
openai>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
